Windows: keep the build warning-free for the code the port touched
The Unix-only helpers in the volume, low-disk and process-ownership code
are now left out of the Windows build instead of being compiled unused,
and pid values are printed as int where the format says so, since pid_t
is 64-bit on Windows.

+++ Workspace/Desktop/Dock/DockIcon.m
@@ -425,7 +425,7 @@
 
   if (pid > 0)
     {
-      [self setToolTip: [NSString stringWithFormat: @"%@ [%d]", appName, pid]];
+      [self setToolTip: [NSString stringWithFormat: @"%@ [%d]", appName, (int)pid]];
 
       /* Register with kernel process monitor so we get notified the
        * instant the process exits, rather than relying on poll-based
